Cambios: + se agregaron muchos DTO + se agregaron DAos + se agrego una conexion a bdd en la nube para que puedan testear en sus casas (funciona muy lento) + se cambio el comportamiento de los combobox, ahora guardan DTOs + se implemento correctamente calcularPremioyDerechos() + se hizo el codigo del estado 3 + se implemento getPolizaDTO() ahora correctamente
PENDIENTE
-Conectar la pantalla faltante
- terminar generarPoliza() de GestorPolizas
- TERMINAR LA PANTALLA 3 DEL CU1
- Conectar la pantalla 3 con la pantalla 1
-Verificar que se cumplan los requisitos funcionales

+++ src/main/java/com/example/aseguradora/CalculadoraMontos.java
@@ -36,7 +36,7 @@ public class CalculadoraMontos{
         premioyDerechosDTO.setDescuento(anio * 10); // Ejemplo: descuento de 10 por cada año de antigüedad
         premioyDerechosDTO.setDerechoEmision(edadConductor * 5); // Ejemplo: derecho de emisión de 5 por cada año de edad
         premioyDerechosDTO.setBaseAnualPrima(premioyDerechosDTO.getPrima() - premioyDerechosDTO.getDescuento());
-        premioyDerechosDTO.setMontoTotal(premioyDerechosDTO.getBaseAnualPrima() + premioyDerechosDTO.getDerechoEmision());
+        premioyDerechosDTO.setMontoTotal((double) (premioyDerechosDTO.getBaseAnualPrima() + premioyDerechosDTO.getDerechoEmision()));
 
         return premioyDerechosDTO;
     }
